[DM-20583] Add TAP client wrapper
This adds two helper functions to work with the catalog.

get_catalog will make a pyvo client for accessing the
catalog with the URL and auth set.

retrieve_query will load a job that was run async and
provide an object to work with that job, such as get
results and delete the job.

These assume they is being run in the LSP notebook
environment, and use environment variables from the
environment to determine the URL to the TAP service,
as well as hook up the auth.

Note this requires new pyvo features.