Replace access token logic
We no longer use JWTs inside the Science Platform and are working
on retiring jupyterhubutils.  Get the access token from the path
where it is mounted in the lab container by default and fall back
on the ACCESS_TOKEN environment variable.  Do not attempt to parse
or validate it (there doesn't seem to be much point just to change
the exception the user will get if it's not valid), but throw an
exception if no token could be found.